Fix ns-3.40 runtime errors and remaining warnings
This commit addresses critical runtime errors and all remaining compilation
warnings to ensure clean execution in ns-3.40.

1. TxErrHeader Trace Source (Runtime Error Fix):
   - In ns-3.40, TxErrHeader trace source has been removed from WiFi MAC
   - Temporarily disabled MAC layer-2 feedback feature (optional)
   - Protocols still work correctly using position-based forwarding
   - Affected: PA-GPSR, GPSR, MM-GPSR (NotifyInterfaceUp/Down)

2. Sign Comparison Warnings:
   - Fixed pagpsr-rst-table.cc: Changed int i to size_t i (2 loops)

3. Unused Variable Warnings:
   - pagpsr-ptable.cc: Removed unused bestFoundDistance and left variables

4. Uninitialized Variable Warnings:
   - pagpsr.cc RecoveryMode: Initialized updated, n_it, NodeRec variables

All three protocols now compile and run cleanly in ns-3.40.

+++ src/pagpsr/model/pagpsr-rst-table.cc
@@ -33,7 +33,7 @@ RstTable::AddEntry (Ipv4Address id, std::pair<std::string, std::pair<uint32_t,Ip
 
       std::vector<std::pair<uint32_t, Ipv4Address>> temp_pvector(pvector.size());
 
-      for (int i=0 ; i < pvector.size() ; i++)
+      for (size_t i = 0; i < pvector.size(); i++)
         temp_pvector[i] = pvector[i].second;
 
       auto it = find(temp_pvector.begin(), temp_pvector.end(), m_pair.second);
@@ -89,7 +89,7 @@ RstTable::FindPacket (Ipv4Address id, std::pair<std::string, std::pair<uint32_t,
 
       std::vector<std::pair<uint32_t, Ipv4Address>> temp_pvector(pvector.size());
 
-      for (int i=0 ; i < pvector.size() ; i++)
+      for (size_t i = 0; i < pvector.size(); i++)
         temp_pvector[i] = pvector[i].second;
 
       auto it = find(temp_pvector.begin(), temp_pvector.end(), m_pair.second);
